trying to fix numpy asserts for shifting/boolean operations

+++ tcnart/core/semantic_type/__init__.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 from dataclasses import dataclass
 from typing import List, Tuple, Dict, Optional, Union
-from numpy import uint64
+from numpy import uint8, uint32, uint64
 
 from . import model
 from .generic import GenericContentConfig
@@ -82,7 +82,7 @@ def _content_storage_to_u64(cfg: ContentTypeConfigStorage) -> uint64:
         return cfg.to_u64()
     if isinstance(cfg, TransformContentConfig):
         return cfg.to_u64()
-    return 0
+    return uint64(0)
 
 
 def _content_storage_to_type(cfg: ContentTypeConfigStorage) -> ContentTypes:
@@ -97,8 +97,8 @@ def _content_storage_to_type(cfg: ContentTypeConfigStorage) -> ContentTypes:
     return ContentTypes.None_
 
 
-def _content_from_type_and_id(ct: ContentTypes, semantic_bits: int) -> ContentTypeConfigStorage:
-    masked = semantic_bits & constants.SEMANTIC_TYPE_MASK
+def _content_from_type_and_id(ct: ContentTypes, semantic_bits: uint32) -> ContentTypeConfigStorage:
+    masked = uint32(semantic_bits) & uint32(constants.SEMANTIC_TYPE_MASK)
     if ct == ContentTypes.Generic:
         return GenericContentConfig.from_u64(masked)
     if ct == ContentTypes.Image:
@@ -119,10 +119,10 @@ class BaseType:
 
     @staticmethod
     def from_identifier(value: BaseIdentifierStorageType) -> "BaseType":
-        st = uint64(uint64(value) & constants.SCALAR_TYPE_MASK) >> constants.SCALAR_TYPE_OFFSET
-        cat = uint64(uint64(value) & constants.CARDINALITY_TYPE_MASK) >> constants.CARDINALITY_TYPE_OFFSET
-        ct = uint64(uint64(value) & constants.CONTAINER_TYPE_MASK) >> constants.CONTAINER_TYPE_OFFSET
-        mrt = uint64(uint64(value) & constants.MEMORY_REPRESENTATION_TYPE_MASK) >> constants.MEMORY_REPRESENTATION_TYPE_OFFSET
+        st = uint64(uint64(value) & uint64(constants.SCALAR_TYPE_MASK)) >> uint8(constants.SCALAR_TYPE_OFFSET)
+        cat = uint64(uint64(value) & uint64(constants.CARDINALITY_TYPE_MASK)) >> uint8(constants.CARDINALITY_TYPE_OFFSET)
+        ct = uint64(uint64(value) & uint64(constants.CONTAINER_TYPE_MASK)) >> uint8(constants.CONTAINER_TYPE_OFFSET)
+        mrt = uint64(uint64(value) & uint64(constants.MEMORY_REPRESENTATION_TYPE_MASK)) >> uint8(constants.MEMORY_REPRESENTATION_TYPE_OFFSET)
         return BaseType(
             scalar_type=ScalarType.from_u8(st) or ScalarType.None_,
             cardinality_type=CardinalityType.from_u8(cat) or CardinalityType.None_,
@@ -140,13 +140,13 @@ class BaseType:
         return BaseType(st, cat, ct, mrt)
 
     def base_type_identifier(self) -> BaseIdentifierStorageType:
-        st_value = uint64(self.scalar_type.as_u8() & 0x3F) << constants.SCALAR_TYPE_OFFSET
-        cat_value = uint64(self.cardinality_type.as_u8() & 0x03) << constants.CARDINALITY_TYPE_OFFSET
-        ct_value = uint64(self.container_type.as_u8() & 0x3F) << constants.CONTAINER_TYPE_OFFSET
-        mrt_value = uint64(self.memory_representation_type.as_u8() & 0x03) << constants.MEMORY_REPRESENTATION_TYPE_OFFSET
-        return int(st_value | cat_value | ct_value | mrt_value)
+        st_value = uint64(self.scalar_type.as_u8() & uint8(0x3F)) << uint8(constants.SCALAR_TYPE_OFFSET)
+        cat_value = uint64(self.cardinality_type.as_u8() & uint8(0x03)) << uint8(constants.CARDINALITY_TYPE_OFFSET)
+        ct_value = uint64(self.container_type.as_u8() & uint8(0x3F)) << uint8(constants.CONTAINER_TYPE_OFFSET)
+        mrt_value = uint64(self.memory_representation_type.as_u8() & uint8(0x03)) << uint8(constants.MEMORY_REPRESENTATION_TYPE_OFFSET)
+        return uint32(st_value | cat_value | ct_value | mrt_value)
 
-    def is_compatible(self, value: int) -> bool:
+    def is_compatible(self, value: uint32) -> bool:
         other = BaseType.from_identifier(value)
 
         def enum_match_query(a, b, match_value) -> bool:
@@ -179,10 +179,10 @@ class SemanticType:
             self.base_type = BaseType()
             self.content_type: ContentTypeConfigStorage = InvalidContentConfig()
         else:
-            base_bits = uint64(semantic_type_id) & constants.BASE_TYPE_MASK
+            base_bits = uint64(semantic_type_id) & uint64(constants.BASE_TYPE_MASK)
             self.base_type = BaseType.from_identifier(base_bits)
             ct = get_content_type(semantic_type_id)
-            self.content_type = _content_from_type_and_id(ct, semantic_type_id)
+            self.content_type = _content_from_type_and_id(ct, uint64(semantic_type_id))
 
     @staticmethod
     def new(semantic_type_id: IdentifierStorageType) -> "SemanticType":
@@ -197,11 +197,11 @@ class SemanticType:
         return SemanticType(value)
 
     def to_identifier(self) -> IdentifierStorageType:
-        content_bits = _content_storage_to_u64(self.content_type) & constants.SEMANTIC_TYPE_MASK
-        base_bits = self.base_type.base_type_identifier() & constants.BASE_TYPE_MASK
+        content_bits = _content_storage_to_u64(self.content_type) & uint64(constants.SEMANTIC_TYPE_MASK)
+        base_bits = self.base_type.base_type_identifier() & uint64(constants.BASE_TYPE_MASK)
         # Ensure content type tag is set consistently with stored config
         content_bits |= set_content_type(_content_storage_to_type(self.content_type))
-        return int((content_bits | base_bits) & 0xFFFFFFFFFFFFFFFF)
+        return uint64((content_bits | base_bits) & uint64(0xFFFFFFFFFFFFFFFF))
 
     # BaseType proxied getters/setters
     def scalar_type(self) -> ScalarType:
@@ -229,7 +229,7 @@ class SemanticType:
         self.base_type.memory_representation_type = memory_representation_type
 
     def set_content_type(self, ct: ContentTypes, config: Optional[int]) -> None:
-        semantic_bits = (config or 0) & constants.SEMANTIC_TYPE_MASK
+        semantic_bits = (config or 0) & uint64(constants.SEMANTIC_TYPE_MASK)
         self.content_type = _content_from_type_and_id(ct, semantic_bits)
 
     def get_content_type(self) -> ContentTypeConfigStorage:
@@ -238,12 +238,12 @@ class SemanticType:
     # Pretty print similar to Rust Debug impl
     def __repr__(self) -> str:
         value = self.to_identifier()
-        base_part = value & constants.BASE_TYPE_MASK
+        base_part = value & uint64(constants.BASE_TYPE_MASK)
 
-        st = (base_part & constants.SCALAR_TYPE_MASK) >> constants.SCALAR_TYPE_OFFSET
-        cat = (base_part & constants.CARDINALITY_TYPE_MASK) >> constants.CARDINALITY_TYPE_OFFSET
-        ct = (base_part & constants.CONTAINER_TYPE_MASK) >> constants.CONTAINER_TYPE_OFFSET
-        mrt = (base_part & constants.MEMORY_REPRESENTATION_TYPE_MASK) >> constants.MEMORY_REPRESENTATION_TYPE_OFFSET
+        st = (base_part & uint64(constants.SCALAR_TYPE_MASK)) >> uint8(constants.SCALAR_TYPE_OFFSET)
+        cat = (base_part & uint64(constants.CARDINALITY_TYPE_MASK)) >> uint8(constants.CARDINALITY_TYPE_OFFSET)
+        ct = (base_part & uint64(constants.CONTAINER_TYPE_MASK)) >> uint8(constants.CONTAINER_TYPE_OFFSET)
+        mrt = (base_part & uint64(constants.MEMORY_REPRESENTATION_TYPE_MASK)) >> uint8(constants.MEMORY_REPRESENTATION_TYPE_OFFSET)
 
         scalar_name = (ScalarType.from_u8(st) or ScalarType.None_).name_str()
         cardinality_name = (CardinalityType.from_u8(cat) or CardinalityType.None_).name_str()
